# Conflicts:
#	Watermelon/Watermelon.vcxproj.filters
#	Watermelon/log.txt
#	Watermelon/src/gameobjects/GameObject.cpp
#	Watermelon/src/gameobjects/GameObject.h
#	Watermelon/src/gameobjects/Player.cpp
#	Watermelon/src/scenes/PlatformerScene.cpp
#	Watermelon/src/scenes/PlatformerScene.h
#	Watermelon/src/scenes/PlatformerScene0.cpp

See previous commit for information on previous commit

+++ Watermelon/src/gameobjects/OrthographicCamera.cpp
@@ -25,6 +25,6 @@ SDL_Rect* OrthographicCamera::GetRect() const {
 }
 
 void OrthographicCamera::Update(const float deltaTime) {
-	cam->x = focus->Position().x;
-	cam->y = focus->Position().y;
+	cam->x = (int)focus->Position().x;
+	cam->y = (int)focus->Position().y;
 }
